% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_version.R
\name{gotm_version}
\alias{gotm_version}
\title{Return the current GOTM model version}
\usage{
gotm_version()
}
\description{
Returns the current version of the GOTM model being used
}
\examples{
print(gotm_version())


}
\author{
Tadhg Moore
}
\keyword{methods}
