% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nml_template_files.R
\name{nml_template_files}
\alias{nml_template_files}
\title{Return the path to a template of current NML files}
\usage{
nml_template_files()
}
\description{
The NML files can change with updated versions of GOTM. This returns a path to a directory with example NML files for running GOTM.
}
\examples{
\dontrun{
nml_template_files()
}

}
\author{
Tadhg Moore
}
\keyword{methods}
