% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_files.R
\name{template_files}
\alias{template_files}
\title{Return the path to a template of current NML files}
\usage{
template_files()
}
\description{
Currently in GOTM 5.4.0 the model can be run with yaml or nml files (old version). This returns a path to a directory with example NML and YAML files for running GOTM.
}
\examples{
\dontrun{
template_files()
}

}
\author{
Tadhg Moore
}
\keyword{methods}
