% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gotm.R
\name{run_gotm}
\alias{run_gotm}
\title{Run the GOTM model}
\usage{
run_gotm(sim_folder = ".", verbose = TRUE, args = character())
}
\arguments{
\item{sim_folder}{the directory where simulation files are contained}

\item{verbose}{should output of GOTM be shown}

\item{args}{Optional arguments to pass to GOTM executable}
}
\description{
This runs the GOTM model on the specific simulation stored in \code{sim_folder}.
The specified \code{sim_folder} must contain a valid NML file.
}
\examples{
sim_folder <- system.file('extdata', package = 'GOTMr')
run_gotm(sim_folder)
}
\author{
Jordan Read, Luke Winslow
}
\keyword{methods}
